exported theme function into its own module
